Refactor: Extract shared enclosure functions to common module
Extract duplicated code from sealed_box.py and ported_box.py into a new
common.py module to reduce code duplication and improve maintainability.

Changes:
- Create src/viberesp/enclosure/common.py with shared functions
- Move calculate_inductance_corner_frequency() to common.py
- Move calculate_hf_rolloff_db() to common.py
- Move calculate_mass_break_frequency() to common.py
- Update sealed_box.py and ported_box.py to import from common

Benefits:
- Reduces code duplication by ~195 lines
- Single source of truth for shared enclosure calculations
- Easier maintenance - future changes only need to be made once
- No behavioral changes - pure refactoring

Validation:
- All 30 sealed box tests pass
- Functions maintain identical functionality
