+++ src/viberesp/enclosure/ported_box.py
@@ -31,220 +31,11 @@ from viberesp.simulation.constants import (
     AIR_DENSITY,
     angular_frequency,
 )
-
-
-def calculate_mass_break_frequency(bl: float, re: float, mms: float) -> float:
-    """
-    Calculate the mass break point frequency.
-
-    Above this frequency, the driver response rolls off at 6 dB/octave
-    due to the motor's inability to accelerate the moving mass.
-
-    Formula from JBL: f_mass = (BL² / Re) / (π × Mms)
-
-    Literature:
-        - Small (1973), "Vented-Box Loudspeaker Systems Part I", JAES
-        - JBL Professional - Tech Note: Characteristics of High-Frequency Compression Drivers
-        - Research: docs/validation/mass_controlled_rolloff_research.md
-
-    Args:
-        bl: Force factor (T·m or N/A)
-        re: DC voice coil resistance (Ω)
-        mms: Moving mass including air load (kg)
-
-    Returns:
-        Mass break point frequency in Hz
-
-    Raises:
-        ValueError: If bl <= 0, re <= 0, or mms <= 0
-
-    Examples:
-        >>> calculate_mass_break_frequency(bl=38.7, re=4.9, mms=0.254)
-        383.1...  # Hz (for B&C 15DS115)
-
-    Validation:
-        Compare with Hornresp mass break point calculation.
-        Expected: <1% deviation from Hornresp values.
-    """
-    if bl <= 0:
-        raise ValueError(f"Force factor BL must be > 0, got {bl} T·m")
-    if re <= 0:
-        raise ValueError(f"DC resistance Re must be > 0, got {re} Ω")
-    if mms <= 0:
-        raise ValueError(f"Moving mass Mms must be > 0, got {mms} kg")
-
-    # JBL formula: f_mass = (BL² / Re) / (π × Mms)
-    # The term BL²/Re represents motor force capability
-    # The term π×Mms represents inertial load
-    # Literature: docs/validation/mass_controlled_rolloff_research.md
-    return (bl ** 2 / re) / (math.pi * mms)
-
-
-def calculate_inductance_corner_frequency(
-    re: float,
-    le: float,
-    frequency: float = None,
-    leach_n: float = 0.6
-) -> float:
-    """
-    Calculate the voice coil inductance corner frequency.
-
-    Above this frequency, the response rolls off at 6 dB/octave
-    due to the voice coil inductance.
-
-    Voice coil inductance is frequency-dependent (semi-inductance).
-    The effective inductance decreases at higher frequencies:
-    Le(f) = Le_DC × (f/1000)^(-n) where n ≈ 0.5-0.7
-
-    This means the corner frequency increases at higher frequencies:
-    f_Le(f) = Re / (2π × Le(f))
-
-    Literature:
-        - Small (1973), "Vented-Box Loudspeaker Systems Part I", JAES
-        - Leach, W.M. "Loudspeaker Voice-Coil Inductance Losses"
-        - Research: docs/validation/mass_controlled_rolloff_research.md
-
-    Args:
-        re: DC voice coil resistance (Ω)
-        le: Voice coil inductance at DC (H)
-        frequency: Frequency in Hz (for frequency-dependent correction)
-        leach_n: Leach exponent (default 0.6, typical range 0.5-0.7)
-
-    Returns:
-        Inductance corner frequency in Hz (returns infinity if Le <= 0)
-
-    Raises:
-        ValueError: If re <= 0
-
-    Examples:
-        >>> # DC corner frequency
-        >>> calculate_inductance_corner_frequency(re=4.9, le=0.0045)
-        173.2...  # Hz (for B&C 15DS115 with 4.5 mH)
-
-        >>> # At 2000 Hz (inductance is less effective)
-        >>> calculate_inductance_corner_frequency(re=4.9, le=0.0045, frequency=2000)
-        275.5...  # Hz (corner frequency shifts higher)
-
-    Validation:
-        Compare with Hornresp inductance corner frequency.
-        Expected: <1% deviation from Hornresp values.
-    """
-    if re <= 0:
-        raise ValueError(f"DC resistance Re must be > 0, got {re} Ω")
-
-    # If inductance is zero or negative, no inductance roll-off
-    if le <= 0:
-        return float('inf')
-
-    # Apply frequency-dependent correction (semi-inductance model)
-    # At higher frequencies, effective inductance decreases
-    # Le(f) = Le_DC × (f/1000)^(-n)
-    # Literature: Leach (2002), voice coil inductance losses
-    if frequency is not None and frequency > 0:
-        # Frequency-dependent inductance
-        # f_ref = 1000 Hz is the reference frequency for Le specification
-        f_ref = 1000.0
-        le_effective = le * (frequency / f_ref) ** (-leach_n)
-
-        # Recalculate corner frequency with effective inductance
-        # f_Le(f) = Re / (2π × Le(f))
-        return re / (2 * math.pi * le_effective)
-
-    # DC corner frequency (no frequency dependence)
-    # f_Le = Re / (2π × Le)
-    # Literature: docs/validation/mass_controlled_rolloff_research.md
-    return re / (2 * math.pi * le)
-
-
-def calculate_hf_rolloff_db(
-    frequency: float,
-    f_le: float,
-    f_mass: float = None,
-    enclosure_type: str = "ported"
-) -> float:
-    """
-    Calculate high-frequency roll-off in dB.
-
-    For direct radiators (ported/sealed boxes):
-        Uses voice coil inductance roll-off (6 dB/octave above f_Le).
-        Optionally applies mass roll-off if f_mass is provided.
-        When f_mass ≈ f_le, creates a second-order filter (12 dB/octave).
-
-    For horn-loaded systems:
-        Uses JBL mass break frequency formula (f_mass from compression driver specs).
-
-    The roll-off is calculated as:
-    - Inductance roll-off: -10·log10(1 + (f/f_Le)²) for first-order low-pass
-    - Mass roll-off: -10·log10(1 + (f/f_mass)²) if f_mass provided
-
-    Literature:
-        - Small (1973), "Vented-Box Loudspeaker Systems Part I", JAES
-        - Leach (2002), "Loudspeaker Voice-Coil Inductance Losses", JAES
-        - Hornresp validation: f_mass ≈ f_le for direct radiators creates correct 2nd-order roll-off
-
-    Args:
-        frequency: Frequency in Hz
-        f_le: Inductance corner frequency in Hz (Re / (2π × Le))
-        f_mass: Mass break point frequency in Hz (optional)
-        enclosure_type: "ported", "sealed", or "horn" (default "ported")
-
-    Returns:
-        Combined roll-off in dB (negative values)
-
-    Raises:
-        ValueError: If frequency <= 0
-
-    Examples:
-        >>> # Ported box (inductance only, first-order)
-        >>> calculate_hf_rolloff_db(frequency=1000, f_le=541, f_mass=None, enclosure_type="ported")
-        -5.3...  # dB
-
-        >>> # Ported box (inductance + mass, second-order when f_mass ≈ f_le)
-        >>> calculate_hf_rolloff_db(frequency=1000, f_le=541, f_mass=500, enclosure_type="ported")
-        -12.8...  # dB (second-order roll-off)
-
-        >>> # Horn-loaded (JBL mass break formula)
-        >>> calculate_hf_rolloff_db(frequency=5000, f_le=173, f_mass=383, enclosure_type="horn")
-        -32.1...  # dB
-
-    Validation:
-        Compare with Hornresp high-frequency response.
-        Expected: Roll-off matches Hornresp within ±2 dB for ported/sealed boxes.
-        BC_18RBX100 validation: f_mass = f_le gives mean error 1.03 dB.
-    """
-    if frequency <= 0:
-        raise ValueError(f"Frequency must be > 0, got {frequency} Hz")
-
-    # Inductance roll-off: first-order low-pass filter
-    # G_le(f) = 1 / (1 + j(f/f_Le))
-    # |G_le(f)|² = 1 / (1 + (f/f_Le)²)
-    # Roll-off in dB = 10·log10(|G_le(f)|²) = -10·log10(1 + (f/f_Le)²)
-    # Literature: Leach (2002), voice coil inductance losses
-    if f_le is not None and f_le < float('inf') and f_le > 0:
-        rolloff_le_db = -10 * math.log10(1 + (frequency / f_le) ** 2)
-    else:
-        rolloff_le_db = 0
-
-    # Mass roll-off behavior depends on enclosure type:
-    # - Horn-loaded: Use JBL compression driver formula (f_mass calculated from BL, Re, Mms)
-    # - Direct radiator: Use f_mass ≈ f_le to create second-order roll-off matching Hornresp
-    #
-    # Validation finding (BC_18RBX100 vs Hornresp):
-    # - Optimal f_mass = 500 Hz vs f_le = 541 Hz (ratio = 0.92)
-    # - Mean error: 1.03 dB when f_mass ≈ f_le
-    # - Creates 12 dB/octave roll-off above ~1 kHz (matches Hornresp)
-    # Literature: Hornresp validation data
-    rolloff_mass_db = 0
-    if f_mass is not None and f_mass > 0:
-        if enclosure_type == "horn":
-            # Horn-loaded compression driver: use JBL mass break formula
-            rolloff_mass_db = -10 * math.log10(1 + (frequency / f_mass) ** 2)
-        elif enclosure_type in ["ported", "sealed"]:
-            # Direct radiator: use f_mass to create second-order roll-off
-            # When f_mass ≈ f_le, this gives 12 dB/octave matching Hornresp
-            rolloff_mass_db = -10 * math.log10(1 + (frequency / f_mass) ** 2)
-
-    return rolloff_le_db + rolloff_mass_db
+from viberesp.enclosure.common import (
+    calculate_inductance_corner_frequency,
+    calculate_hf_rolloff_db,
+    calculate_mass_break_frequency,
+)
 
 
 @dataclass
